+++ app/views/articles/show.html.haml
@@ -32,3 +32,5 @@
 .container
   = link_to new_article_comment_path(@article) do
     %p.btn-secondary  コメントを追加
+
+= javascript_pack_tag 'article'
